# shared/api-contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_image_job_api_jobs_images_post": {
        "properties": {
          "files": {
            "description": "One or more image files or ZIP archives. Supported extensions: .jpg, .jpeg, .png, .webp, .zip.",
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_create_image_job_api_jobs_images_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "seo-studio maintainers"
    },
    "description": "\nThe seo-studio API powers the local proof-of-concept workflow for image optimization,\nwebsite crawling, AI metadata generation, and exports.\n\nThe POC uses local file storage, a separate local Ollama runtime, and Supabase\nAuth for dashboard access. Long-running worker queues and persistent job\ndatabases are intentionally deferred until beta readiness.\n",
    "summary": "Local API for the seo-studio optimization POC.",
    "title": "seo-studio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs/images": {
      "post": {
        "description": "Creates an image job from one or more uploaded files. Accepts direct `.jpg`, `.jpeg`, `.png`, and `.webp` images plus `.zip` archives that contain supported images. Direct image files and ZIP entries are limited to 5MB each. ZIP archives are safely inspected to prevent path traversal.",
        "operationId": "create_image_job_api_jobs_images_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_image_job_api_jobs_images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageJobCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The upload is empty, unsupported, too large, unsafe, or not a valid image."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create image upload job",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "description": "Returns status for an image job created by the upload endpoint, including the number of validated uploaded image files.",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get job status",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/brand-context": {
      "get": {
        "description": "Returns brand context documents attached to an image job and the extracted plain text that will be included in AI image metadata prompts. Empty documents and text are returned when no brand context has been uploaded.",
        "operationId": "read_brand_context_api_jobs__job_id__brand_context_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read brand context documents",
        "tags": [
          "image jobs"
        ]
      },
      "post": {
        "description": "Attaches one or more brand context documents to an image job. Supports `.txt`, `.docx`, and `.pdf` files. The backend stores the original documents under the job folder, extracts plain text, truncates the combined context to the configured prompt limit, and persists it in `storage/uploads/{job_id}/job.json` for later AI metadata generation.",
        "operationId": "upload_brand_context_api_jobs__job_id__brand_context_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_brand_context_api_jobs__job_id__brand_context_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "The upload is empty, unsupported, too large, or has no extractable text."
          },
          "404": {
            "description": "Job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload brand context documents",
        "tags": [
          "image jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/files": {
      "get": {
        "description": "Returns validated direct image uploads and extracted ZIP image entries for an image job.",
        "operationId": "list_job_files_api_jobs__job_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobFileListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List uploaded image job files",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/images/accept-all": {
      "post": {
        "description": "Marks multiple persisted metadata rows as accepted. The request body contains the uploaded image ids to approve.",
        "operationId": "accept_all_image_metadata_api_jobs__job_id__images_accept_all_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageMetadataBulkAcceptRequest",
                "description": "Uploaded image file ids whose metadata rows should be marked accepted."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadataListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "One or more rows cannot be approved in their current state."
          },
          "404": {
            "description": "Job or one or more metadata rows were not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve multiple metadata rows",
        "tags": [
          "image jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/images/metadata": {
      "get": {
        "description": "Returns generated AI metadata for an image job. Empty results are returned when metadata has not been generated yet.",
        "operationId": "list_image_metadata_api_jobs__job_id__images_metadata_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadataListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List AI image metadata",
        "tags": [
          "image jobs"
        ]
      },
      "post": {
        "description": "Generates filename suggestions, alt text, captions, and confidence scores for every image in a job. Images are processed sequentially through the configured AI provider. Failed rows are returned with failed status while successful rows are persisted for review.",
        "operationId": "generate_all_image_metadata_api_jobs__job_id__images_metadata_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadataListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Job not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate AI metadata for all image files",
        "tags": [
          "image jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/processed/{filename}": {
      "get": {
        "description": "Downloads one processed image from `storage/processed/{job_id}/images`. The filename must be a direct processed filename returned by the compression result.",
        "operationId": "download_processed_image_api_jobs__job_id__processed__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Filename is invalid."
          },
          "404": {
            "description": "Processed file not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download processed image",
        "tags": [
          "image jobs"
        ]
      }
    },
...
  },
...
}